+++ src/build-tools/pyiodide-docutils.ts
@@ -4,6 +4,8 @@
 import { loadPyodide, type PyodideAPI } from 'pyodide'
 import type { PyProxy, PyProxyWithGet } from 'pyodide/ffi'
 
+import directiveCode from './directives.py?raw'
+
 export async function load(): Promise<{
 	pyodide: PyodideAPI
 	core: PyProxy
@@ -12,26 +14,8 @@ export async function load(): Promise<{
 		packages: ['docutils'],
 		packageBaseUrl: 'https://cdn.jsdelivr.net/pyodide/v0.28.1/full/',
 	})
-
-	await pyodide.runPython(`
-		from docutils.nodes import Element, General
-		from docutils.parsers import rst
-		
-		class PlotlyElement(General, Element):
-			tagname = 'plotly'
-				
-		class PlotlyDirective(rst.Directive):
-			required_arguments = 1
-			option_spec = dict(href=rst.directives.uri)
-
-			def run(self):
-				node = PlotlyElement(self.content)
-				node.source, node.line = self.state_machine.get_source_and_line(self.lineno)
-		
-		rst.directives.register_directive("plotly", PlotlyDirective)
-	`)
-
 	const core = await pyodide.pyimport('docutils.core')
+	await pyodide.runPython(directiveCode)
 	return { pyodide, core }
 }
